First beta of suid-wrapper
This is the initial version of suid-wrapper.
The program works in three steps: check, grant, and execute.

This version comprises the following checks:
 * Filter the UID/GID allowed to execute the wrapper
 * Filter the path from which the wrapper may be executed
 * Filter the minimum and maximum number of arguments that
   may be passed to the program
 * Filter arguments that may be passed to the target using
   regular expressions

... the following grants:
 * Drop capabilities of the caller to only give the ones that
   are stritly necessary
 * Change the real UID/GID of the caller

... the following execution options:
 * Pass arguments to the program (prior to caller arguments)

Each configuration makes a single purpose wrapper. Multiple
purposes means multiple wrappers, after compilation, the
binary is given desired permissions (setuid or setgid bits,
owner and group) to be able to grant the appropriate privileges.

Signed-off-by: Emmanuel Thierry <emmanuel.thierry@sekil.fr>